refactor: build JetBrains feed OkHttp client inline, share proxy setup
CoderHttpClientBuilder is for Coder endpoints — having it grow a
"without TLS alternate hostname" variant just to serve the JetBrains
data services feed was a smell. Move the JetBrains client construction
into IdeFeedManager where it belongs, with only what it actually needs:
user-agent, logging, proxy, retry. No Coder TLS context or hostname
verifier touches that traffic anymore.
